movietic/movieticket/requirements.txt
Django==5.2.9
djangorestframework==3.15.1
psycopg[binary]==3.1.18
pymongo==4.7.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.0
